requirements.txt
Django==5.0.4
django-widget-tweaks~=1.5.0
django-allauth~=0.61.1
django-environ==0.11.2
dj-database-url==2.1.0
python-decouple==3.8
line-bot-sdk==3.11.0
psycopg2==2.9.9
gunicorn==22.0.0
whitenoise==6.6.0
django-widget-tweaks~=1.5.0
hashids==1.3.1
dj-static==0.0.6
Pillow==10.3.0
